Enable ex_ready when LSU is working on a non-split XIF req.
According to the comment disabling the ready explicitly in such condition,
  // XIF transactions take precedence, thus the LSU is not ready for EX in case of an XIF request.
However, in my view this comment on XIF transactions is misleading, as it causes the
pipeline to stall for no reason (lsu_split_0_o is anyways low, and xif_req is
like any other request).
This caused a spurious stall on two back-to-back XIF loads, which then
broke our test pipeline.